+++ attributes/epel-next-testing-debuginfo.rb
@@ -1,10 +1,10 @@
 default['yum']['epel-next-testing-debuginfo']['repositoryid'] = 'epel-next-testing-debuginfo'
 default['yum']['epel-next-testing-debuginfo']['description'] =
-  "Extra Packages for #{node['platform_version'].to_i} - $basearch - Next - Testing Debug"
+  "Extra Packages for #{yum_epel_release} - $basearch - Next - Testing Debug"
 default['yum']['epel-next-testing-debuginfo']['mirrorlist'] =
-  "https://mirrors.fedoraproject.org/mirrorlist?repo=epel-testing-next-debug-#{node['platform_version'].to_i}&arch=$basearch"
+  "https://mirrors.fedoraproject.org/mirrorlist?repo=epel-testing-next-debug-#{yum_epel_release}&arch=$basearch"
 default['yum']['epel-next-testing-debuginfo']['gpgkey'] =
-  "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{node['platform_version'].to_i}"
+  "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{yum_epel_release}"
 default['yum']['epel-next-testing-debuginfo']['gpgcheck'] = true
 default['yum']['epel-next-testing-debuginfo']['enabled'] = false
 default['yum']['epel-next-testing-debuginfo']['managed'] = false