+++ attributes/epel-next-testing-source.rb
@@ -1,10 +1,10 @@
 default['yum']['epel-next-testing-source']['repositoryid'] = 'epel-next-testing-source'
 default['yum']['epel-next-testing-source']['description'] =
-  "Extra Packages for #{node['platform_version'].to_i} - $basearch - Next - Testing Source"
+  "Extra Packages for #{yum_epel_release} - $basearch - Next - Testing Source"
 default['yum']['epel-next-testing-source']['mirrorlist'] =
-  "https://mirrors.fedoraproject.org/mirrorlist?repo=testing-source-epel#{node['platform_version'].to_i}&arch=$basearch"
+  "https://mirrors.fedoraproject.org/mirrorlist?repo=testing-source-epel#{yum_epel_release}&arch=$basearch"
 default['yum']['epel-next-testing-source']['gpgkey'] =
-  "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{node['platform_version'].to_i}"
+  "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{yum_epel_release}"
 default['yum']['epel-next-testing-source']['gpgcheck'] = true
 default['yum']['epel-next-testing-source']['enabled'] = false
 default['yum']['epel-next-testing-source']['managed'] = false