+++ attributes/epel-source.rb
@@ -1,14 +1,10 @@
 default['yum']['epel-source']['repositoryid'] = 'epel-source'
-
-if platform?('amazon')
-  default['yum']['epel-source']['description'] = 'Extra Packages for 7 - $basearch - Source'
-  default['yum']['epel-source']['mirrorlist'] = 'https://mirrors.fedoraproject.org/mirrorlist?repo=epel-source-7&arch=$basearch'
-  default['yum']['epel-source']['gpgkey'] = 'https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-7'
-else
-  default['yum']['epel-source']['description'] = "Extra Packages for #{node['platform_version'].to_i} - $basearch - Source"
-  default['yum']['epel-source']['mirrorlist'] = "https://mirrors.fedoraproject.org/mirrorlist?repo=epel-source-#{node['platform_version'].to_i}&arch=$basearch"
-  default['yum']['epel-source']['gpgkey'] = "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{node['platform_version'].to_i}"
-end
+default['yum']['epel-source']['description'] =
+  "Extra Packages for #{yum_epel_release} - $basearch - Source"
+default['yum']['epel-source']['mirrorlist'] =
+  "https://mirrors.fedoraproject.org/mirrorlist?repo=epel-source-#{yum_epel_release}&arch=$basearch"
+default['yum']['epel-source']['gpgkey'] =
+  "https://download.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-#{yum_epel_release}"
 default['yum']['epel-source']['gpgcheck'] = true
 default['yum']['epel-source']['enabled'] = false
 default['yum']['epel-source']['managed'] = false